Permissions to edit checked
It is no longer possible to edit another user's vote

+++ src/delFromVote.php
@@ -4,19 +4,34 @@
 
 	$vote_id = $_GET["vote_id"];
 	$option_id = $_GET["option_id"];
+	$canWrite = true;
 
-	$queryResult = queryDatabase( 
-		$dbConnect,
-		"delete from vote_options ".
-		"where option_id = $1",
-		array( $option_id )
-	);
-
-	if( isset( $queryResult ) && !is_object($queryResult) )
+	if( !$actUser['administrator'] )
 	{
-		header("Location: voteEdit.php?vote_id=" . $vote_id );
-		exit();
+		$vote = getVote($dbConnect,$vote_id);
+		if( $vote["user_id"] != $actUser['id'] )
+		{
+			$canWrite = false;
+		}
 	}
+
+	if( $canWrite )
+	{
+		$queryResult = queryDatabase( 
+			$dbConnect,
+			"delete from vote_options ".
+			"where option_id = $1",
+			array( $option_id )
+		);
+	
+		if( isset( $queryResult ) && !is_object($queryResult) )
+		{
+			header("Location: voteEdit.php?vote_id=" . $vote_id );
+			exit();
+		}
+	}
+	else
+		$error = "Keine Berechtigung!";
 ?>
 
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Strict//EN">
@@ -32,7 +47,7 @@
 		<?php
 			include( "includes/components/headerlines.php" );
 
-			include "../includes/components/error.php";
+			include "includes/components/error.php";
 		?>
 		<?php include( "includes/components/footerlines.php" ); ?>
 	</body>
